Correctly apply secrets to config values
